signup, login, private route

+++ routes/profile.js
@@ -34,10 +34,19 @@ async function routes(fastify, options) {
           "5xx": { ...BasicMessage, description: "Failed response" },
         },
         body: BasicItem,
+        security: [
+          {
+            Bearer: [],
+          },
+        ],
       },
+      preValidation: [fastify.authenticate],
     },
     async (req, reply) => {
       try {
+        const { email } = req.user;
+        if (email !== "admin@admin.com") throw Error("wey, anda bukan admin");
+
         const {
           filter,
           imgSrc,
@@ -74,10 +83,19 @@ async function routes(fastify, options) {
         },
         body: BasicItem,
         params: BasicId,
+        security: [
+          {
+            Bearer: [],
+          },
+        ],
       },
+      preValidation: [fastify.authenticate],
     },
     async (req, reply) => {
       try {
+        const { email } = req.user;
+        if (email !== "admin@admin.com") throw Error("wey, anda bukan admin");
+
         const {
           filter,
           imgSrc,
@@ -135,11 +153,20 @@ async function routes(fastify, options) {
           "2xx": { ...BasicMessage, description: "Successful item deletion" },
           "5xx": { ...BasicMessage, description: "Failed response" },
         },
+        security: [
+          {
+            Bearer: [],
+          },
+        ],
         params: BasicId,
       },
+      preValidation: [fastify.authenticate],
     },
     async (req, reply) => {
       try {
+        const { email } = req.user;
+        if (email !== "admin@admin.com") throw Error("wey, anda bukan admin");
+
         const returnVal = await fastify.pg.query(
           `DELETE FROM profiles WHERE id=$1;`,
           [req.params.id]
